[Determinator] Migrate Replay V8 hooks to v147 API
## Problems
- Replay-modified V8 source no longer compiled against v147.
  - Many V8 APIs changed: free predicate functions, `Cast<T>`, `Tagged<T>`, `DirectHandle<T>`.
  - `Execution::Call`, `JsonStringify`, `JsonParser::Parse`, `CommonFrame::Summarize`, `IncrementalStringBuilder::Finish` got new signatures.
  - Renamed members/flags: `JSDate::value`, `Script::Type::kWasm`, `v8_flags.minor_gc_task`, `v8_flags.incremental_marking_task`.

## Solutions
- Updated Replay code/diagnostics call sites to the v147 API.
  - Switched member predicates/accessors to free functions and `Cast<T>`.
  - Re-typed Replay helper signatures and callers to `Tagged<Script>` / `DirectHandle<Object>` consistently across the tree.
  - Adapted to new return/argument types with the necessary casts and `base::VectorOf`/empty-vector args.
- No Replay interventions added or removed; behavior is unchanged.

+++ src/objects/value-serializer.cc
@@ -589,9 +589,10 @@ void ValueSerializer::WriteString(DirectHandle<String> string) {
     // as two bytes to ensure serialized buffers have a consistent size.
     if (recordreplay::IsRecordingOrReplaying("values", "ValueSerializer::WriteString")) {
       auto new_chars = base::OwnedVector<base::uc16>::NewForOverwrite(chars.length());
-      for (size_t i = 0; i < chars.length(); i++)
+      for (int i = 0; i < chars.length(); i++)
         new_chars[i] = chars[i];
-      uint32_t byte_length = new_chars.size() * sizeof(base::uc16);
+      uint32_t byte_length =
+          static_cast<uint32_t>(new_chars.size() * sizeof(base::uc16));
       // The existing reading code expects 16-byte strings to be aligned.
       if ((buffer_size_ + 1 + BytesNeededForVarint(byte_length)) & 1)
         WriteTag(SerializationTag::kPadding);
